requirements.txt
azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.20.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
numpy==2.0.0